fix: shared validateStatusExpiration for account profile and edit status modal
