+++ apps/meteor/client/lib/statusDurations.ts
@@ -1,3 +1,5 @@
+import type { TranslationContextValue } from '@rocket.chat/ui-contexts';
+
 type DurationOption = {
 	value: string;
 	labelKey: 'Status_dont_clear' | 'Status_30_minutes' | 'Status_1_hour' | 'Status_choose_date_and_time';
@@ -28,3 +30,29 @@ export const STATUS_DURATION_OPTIONS: DurationOption[] = [
 		},
 	},
 ];
+
+export const validateStatusExpiration = (
+	value: string,
+	{ statusCustomDate, statusCustomTime }: { statusCustomDate?: string; statusCustomTime?: string },
+	t: TranslationContextValue['translate'],
+): string | true => {
+	if (value !== 'custom') {
+		return true;
+	}
+
+	const expiresAt = STATUS_DURATION_OPTIONS.find((o) => o.value === value)?.getExpiresAt?.({
+		now: new Date(),
+		customDate: statusCustomDate,
+		customTime: statusCustomTime,
+	});
+
+	if (!expiresAt) {
+		return t('Status_choose_date_and_time');
+	}
+
+	if (expiresAt <= new Date()) {
+		return t('Status_expiration_must_be_future');
+	}
+
+	return true;
+};
